Integrated walking animation with camera bound doggo
