--- projectFiles/Lost and Found/Assets/Code/Controller/MAOPlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a65c5660ca122414cb4c3fe569eae395
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
